IGNITE-15406 Ignite, Ignition, and IgniteManager moved to org.apache.ignite package
