+++ modules/cli/src/integrationTest/java/org/apache/ignite/cli/ITConfigCommandTest.java
@@ -24,7 +24,7 @@ import java.net.ServerSocket;
 import java.nio.file.Path;
 import io.micronaut.context.ApplicationContext;
 import io.micronaut.context.env.Environment;
-import org.apache.ignite.app.IgnitionManager;
+import org.apache.ignite.IgnitionManager;
 import org.apache.ignite.cli.spec.IgniteCliSpec;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
