+++ modules/client/src/test/java/org/apache/ignite/client/fakes/FakeIgnite.java
@@ -17,7 +17,7 @@
 
 package org.apache.ignite.client.fakes;
 
-import org.apache.ignite.app.Ignite;
+import org.apache.ignite.Ignite;
 import org.apache.ignite.internal.processors.query.calcite.QueryProcessor;
 import org.apache.ignite.table.manager.IgniteTables;
 import org.apache.ignite.tx.IgniteTransactions;
